# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "manypost API",
    "version": "0.0.1",
    "description": "API do manypost — agendamento/publicação multicanal. Humanos usam sessão Clerk; máquinas usam `Authorization: Bearer mp_live_…`. Erros seguem problem+json (RFC 9457): o campo `title` carrega o código estável do erro."
  },
  "servers": [
    {
      "url": "/",
      "description": "mesma origem (proxy do Next)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "clerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "sessão Clerk humana — Authorization: Bearer <token>"
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key mp_live_ para superfícies de máquina — Authorization: Bearer <chave>"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "about:blank"
          },
          "title": {
            "type": "string",
            "description": "código estável do erro",
            "example": "common.not_found"
          },
          "status": {
            "type": "integer",
            "example": 404
          },
          "detail": {
            "type": "string",
            "example": "post não encontrado"
          },
          "extra": {
            "description": "campos extras (ex.: issues de validação)"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "Media": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "URL pública (as redes baixam por aqui)"
          },
          "mime": {
            "type": "string",
            "example": "image/png"
          },
          "source": {
            "type": "string",
            "description": "`ai` = gerada por IA; `upload` = enviada por alguém"
          },
          "byteSize": {
            "type": "integer"
          },
          "width": {
            "type": [
              "integer",
              "null"
            ]
          },
          "height": {
            "type": [
              "integer",
              "null"
            ]
          },
          "alt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "url",
          "mime",
          "source",
          "byteSize",
          "width",
          "height",
          "alt",
          "createdAt"
        ]
      },
...
      "PubMedia": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "mime": {
            "type": "string",
            "example": "image/png"
          },
          "source": {
            "type": "string",
            "description": "`ai` = gerada por IA; `upload` = enviada por alguém"
          },
          "byteSize": {
            "type": "integer"
          },
          "width": {
            "type": [
              "integer",
              "null"
            ]
          },
          "height": {
            "type": [
              "integer",
              "null"
            ]
          },
          "alt": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "url",
          "mime",
          "source",
          "byteSize",
          "width",
          "height",
          "alt",
          "createdAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/v1/media/upload": {
      "post": {
        "tags": [
          "media"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Envia mídia (multipart) — MIME real por magic bytes",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "arquivo (MIME real detectado por magic bytes)"
                  },
                  "alt": {
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "mídia criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media/from-url": {
      "post": {
        "tags": [
          "media"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Importa mídia por URL (anti-SSRF, re-valida a cada redirect)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "alt": {
                    "type": "string",
                    "maxLength": 1500
                  }
                },
                "required": [
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "mídia criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media": {
      "get": {
        "tags": [
          "media"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Lista a biblioteca de mídia",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "50"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "itens de mídia",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media/{id}": {
      "patch": {
        "tags": [
          "media"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Atualiza o texto alternativo (alt) de uma mídia",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "alt": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 1500
                  }
                },
                "required": [
                  "alt"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "alt atualizado"
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "media"
        ],
        "security": [
          {
            "clerkAuth": []
          }
        ],
        "summary": "Remove uma mídia (soft — arquivo fica p/ posts já agendados)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "mídia removida"
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/v1/media/upload": {
      "post": {
        "tags": [
          "public-media"
        ],
        "summary": "Envia mídia multipart — MIME real por magic bytes (escopo media:write)",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "alt": {
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "mídia criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PubMedia"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate limit — aguarde e tente de novo",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/public/v1/media/from-url": {
      "post": {
        "tags": [
          "public-media"
        ],
        "summary": "Importa mídia por URL (anti-SSRF) (escopo media:write)",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "alt": {
                    "type": "string",
                    "maxLength": 1500
                  }
                },
                "required": [
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "mídia criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PubMedia"
                }
              }
            }
          },
          "400": {
            "description": "requisição fora do contrato (problem+json)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate limit — aguarde e tente de novo",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/public/v1/media": {
      "get": {
        "tags": [
          "public-media"
        ],
        "summary": "Lista a biblioteca de mídia (escopo media:write)",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "50"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "itens de mídia",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PubMedia"
                  }
                }
              }
            }
          },
          "401": {
            "description": "não autenticado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "papel insuficiente (requer ADMIN/OWNER)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "rate limit — aguarde e tente de novo",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}